refactored for user name tests
